Move the RepoMapData class to a common library
